Standardize all email templates to a reusable blue layout
Unify every email in src/mails under the htmlContentCommentCards
style (Segoe UI, #004aad accent, standard Rolf Wittmer footer) via a
single reusable mailLayout() function. Rename mail files/classes to
English for consistency and update imports without breaking call sites.
