--- src/mails/mailsOrder.js
@@ -1,180 +1,0 @@
-require('dotenv').config();
-
-class MailsOrder {
-  static htmlNewOrder(company) {
-    const htmlContent = `
-          <!DOCTYPE html>
-          <html lang="es">
-          <head>
-              <meta charset="UTF-8">
-              <meta name="viewport" content="width=device-width, initial-scale=1.0">
-              <style>
-                  body {
-                      font-family: arial, sans-serif;
-                  }
-                  .container-pass {
-                      max-width: 600px;
-                      margin: 0 auto;
-                  }
-                  .image-container {
-                      text-align: center;
-                      margin-top: 20px;
-                  }
-                  .styled-image {
-                    width: 230px;
-                  }
-                  .credencial-container {
-                    display: flex;
-                    align-items: center;
-                  }
-                  .credencial-container p, h3 {
-                    margin: 0px;
-                  }
-                  .credential-tittle {
-                    margin-right: 10px;
-                  }
-                  .button {
-                      display: inline-block;
-                      padding: 10px 20px;
-                      text-decoration: none;
-                      background-color: #F29100;
-                      border-radius: 150px;
-                      color: #ffffff;
-                  }
-                  .buttom-container {
-                      display: flex;
-                      justify-content: center;
-                      margin-bottom: 20px;
-                      margin-top: 20px;
-                  }
-                  .body-container p {
-                    margin: 0px;
-                  }
-                  .body-container {
-                    margin-bottom: 20px;
-                  }
-                  .footer p,h3 {
-                    margin: 0px;
-                  }
-                  .footer {
-                    margin-top: 20px;
-                  }
-              </style>
-          </head>
-          <body>
-              <div class="container-pass">
-                  <div class="image-container">
-                    <img src="https://reservation.rwittmer.com/logo_rwittmer.png" alt="Imagen con Estilos" class="styled-image">
-                  </div>
-                  <h2>Has recibido un nuevo pedido</h2>
-                  <div class="body-container">
-                    <p>Hola has recibido un nuevo pedido para la empresa</p>
-                    <p>${company}</p>
-                  </div>
-                   <div class="body-container">
-                    <p>Inicia sesión con tus credenciales para revisarlo</p>
-                  </div>
-                  <div class="buttom-container">
-                      <a href=${process.env.URL_STAFFS} class="button">Iniciar sesión</a>
-                  </div>
-                  <div class="footer">
-                    <p>Atentamente</p>
-                    <h3>Rolf Wittmer</h3>
-                  </div>
-              </div>
-          </body>
-          </html>
-          `;
-    return htmlContent;
-  }
-
-  static htmlNewRequest(company) {
-    const htmlContent = `
-          <!DOCTYPE html>
-          <html lang="es">
-          <head>
-              <meta charset="UTF-8">
-              <meta name="viewport" content="width=device-width, initial-scale=1.0">
-              <style>
-                  body {
-                      font-family: arial, sans-serif;
-                  }
-                  .container-pass {
-                      max-width: 600px;
-                      margin: 0 auto;
-                  }
-                  .image-container {
-                      text-align: center;
-                      margin-top: 20px;
-                  }
-                  .styled-image {
-                    width: 230px;
-                  }
-                  .credencial-container {
-                    display: flex;
-                    align-items: center;
-                  }
-                  .credencial-container p, h3 {
-                    margin: 0px;
-                  }
-                  .credential-tittle {
-                    margin-right: 10px;
-                  }
-                  .button {
-                      display: inline-block;
-                      padding: 10px 20px;
-                      text-decoration: none;
-                      background-color: #F29100;
-                      border-radius: 150px;
-                      color: #ffffff;
-                  }
-                  .buttom-container {
-                      display: flex;
-                      justify-content: center;
-                      margin-bottom: 20px;
-                      margin-top: 20px;
-                  }
-                  .body-container p {
-                    margin: 0px;
-                  }
-                  .body-container {
-                    margin-bottom: 20px;
-                  }
-                  .footer p,h3 {
-                    margin: 0px;
-                  }
-                  .footer {
-                    margin-top: 20px;
-                  }
-              </style>
-          </head>
-          <body>
-              <div class="container-pass">
-                  <div class="image-container">
-                    <img src="https://reservation.rwittmer.com/logo_rwittmer.png" alt="Imagen con Estilos" class="styled-image">
-                  </div>
-                  <h2>Has recibido un nuevo pedido</h2>
-                  <div class="body-container">
-                    <p>Hola has recibido un nuevo requerimiento para la embarcación</p>
-                    <p>${company}</p>
-                  </div>
-                   <div class="body-container">
-                    <p>Inicia sesión con tus credenciales para revisarlo</p>
-                  </div>
-                  <div class="buttom-container">
-                      <a href=${process.env.URL_STAFFS} class="button">Iniciar sesión</a>
-                  </div>
-                  <div class="footer">
-                    <p>Atentamente</p>
-                    <h3>Rolf Wittmer</h3>
-                  </div>
-              </div>
-          </body>
-          </html>
-          `;
-    return htmlContent;
-  }
-
-}
-
-module.exports =  MailsOrder;
